+++ src/Vl1.js
@@ -1,13 +1,8 @@
 import correspondance from './correspondance-icônes.yaml'
 import vélo from '../vélos/1.yaml'
+import {Card} from './ui'
+import {domain} from './utils'
 
-const domain = (url) => {
-	const a = document.createElement('a')
-	a.href = url
-	const res = a.hostname.replace('www.', '')
-
-	return res === 'localhost' ? null : res
-}
 
 export default ({}) => (
 	<div>
@@ -43,22 +38,11 @@ export default ({}) => (
 									{composant}
 								</div>
 							</div>
-							<div css={`
-box-shadow: 0 1px 3px rgba(41, 117, 209, 0.12), 0 1px 2px rgba(41, 117, 209, 0.24);
-border: 4px solid var(--lighterColor);
-    border-radius: 0.3rem;
-	padding: 1rem;
-
-	:hover {
-box-shadow: 0px 2px 4px -1px rgba(41, 117, 209, 0.2), 0px 4px 5px 0px rgba(41, 117, 209, 0.14), 0px 1px 10px 0px rgba(41, 117, 209, 0.12);
-	}
-	
-
-								`}><div>
-									<span css="font-size: 90%; font-weight: bold; margin-right: .4rem">{marque}</span>
-									<span>{modèle}</span></div>
+							<Card><div>
+								<span css="font-size: 90%; font-weight: bold; margin-right: .4rem">{marque}</span>
+								<span>{modèle}</span></div>
 								<div>{prix} {domain(url) && <span>sur <a href={url}>{domain(url)}</a></span>}</div>
-							</div>
+							</Card>
 						</li>
 					)
 				)}
